test-generator/generate-test.pl: Tool to generate tests
This script generates different random variants of tests.

It selects random questions.
Then selects random position for the right answer.
Finally collects the answers from the DB and populates the questions hash.

After all of the above is done it generates two PDFs. One is the actual test and the other is the key for checking the test :)